refactor : SceneDelegate remove
