+++ Terminal-iOS/AppDelegate.swift
@@ -14,8 +14,27 @@ import CoreData
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate, UNUserNotificationCenterDelegate {
 
+    var window: UIWindow?
+    
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
+        window = UIWindow()
+        let home = HomeView()
+
+        // 리프레쉬 토큰이 없으면 -> 로그인
+        if KeychainWrapper.standard.string(forKey: "refreshToken") == nil {
+            KeychainWrapper.standard.set("temp", forKey: "accessToken")
+            let howView = UINavigationController(rootViewController: home)
+            window?.rootViewController = howView
+        } else {
+            print("토큰이 유효합니다..")
+            print("로그인 완료")
+            print("accessToken : ", KeychainWrapper.standard.string(forKey: "accessToken")!)
+            let main = ViewController()
+            window?.rootViewController = main
+        }
+
+        window?.makeKeyAndVisible()
         
         let center = UNUserNotificationCenter.current()
         center.delegate = self
@@ -49,20 +68,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate, UNUserNotificationCenterD
         let pushToken = KeychainWrapper.standard.set(deviceTokenString, forKey: "pushToken")
         print("pushToken 성공여부:",pushToken)
         print("pushToken:",KeychainWrapper.standard.string(forKey: "pushToken")!)
-    }
-    
-    // MARK: UISceneSession Lifecycle
-
-    func application(_ application: UIApplication, configurationForConnecting connectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
-        // Called when a new scene session is being created.
-        // Use this method to select a configuration to create the new scene with.
-        return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
-    }
-
-    func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {
-        // Called when the user discards a scene session.
-        // If any sessions were discarded while the application was not running, this will be called shortly after application:didFinishLaunchingWithOptions.
-        // Use this method to release any resources that were specific to the discarded scenes, as they will not return.
     }
     
     // MARK: - Core Data stack
